--- Lab07_Agent的守护脚本/watch_dog.py
@@ -1,6 +1,0 @@
-import time
-
-with open('testquit.log') as f:
-    text = f.read()
-if 'Error' in text:
-    print('请立即修复！')
